Fix ganache deploying with parameters from estimation
