+++ package.json
@@ -26,7 +26,7 @@
   "license": "(Apache-2.0 OR LGPL-2.1-only)",
   "devDependencies": {
     "coveralls": "^3.0.0",
-    "ganache-cli": "^6.1.0-beta.4",
+    "ganache-cli": "^6.1.6",
     "lcov-result-merger": "^2.0.0",
     "lerna": "^2.9.0",
     "ncp": "^2.0.0",
