+++ packages/dapp/package.json
@@ -33,7 +33,7 @@
     "reselect": "^3.0.1",
     "rxjs": "^5.5.7",
     "sanitize-html": "^1.18.2",
-    "semantic-ui-react": "^0.79.1",
+    "semantic-ui-react": "~0.82.0",
     "sort-by": "^1.2.0",
     "store": "^2.0.12",
     "styled-components": "^3.2.5"
