+++ packages/dapp/src/reducers/newsrooms.ts
@@ -1,6 +1,5 @@
 import { Set } from "immutable";
 import { AnyAction } from "redux";
-import { NewsroomWrapper, EthAddress } from "@joincivil/core";
 import { newsroomActions } from "../actionCreators/newsrooms";
 
 export function currentUserNewsrooms(state: Set<string> = Set<string>(), action: AnyAction): Set<string> {
