+++ packages/ethapi/src/ethapi.ts
@@ -16,31 +16,11 @@ import {
 import { AbiDecoder } from "./abidecoder";
 import { requireAccount } from "./helpers";
 
-const DEFAULT_HTTP_NODE = "http://localhost:8545";
-
 const debug = Debug("civil:ethapi");
 
 const POLL_MILLISECONDS = 1000;
 
 export class EthApi {
-  public static detectProvider(onAccountSet?: () => void): Web3.Provider {
-    let provider: Web3.Provider;
-    // Try to use the window's injected provider
-    if (EthApi.hasInjectedProvider()) {
-      const injectedWeb3: Web3 = (window as any).web3;
-      provider = injectedWeb3.currentProvider;
-      debug("Using injected web3 provider");
-    } else {
-      // TODO(ritave): Research using Infura
-      provider = new Web3.providers.HttpProvider(DEFAULT_HTTP_NODE);
-      debug("No web3 provider provided or found injected, defaulting to HttpProvider");
-    }
-    return provider;
-  }
-
-  public static hasInjectedProvider(): boolean {
-    return typeof window !== "undefined" && (window as any).web3 !== undefined;
-  }
   // Initialized for sure by the helper method setProvider used in constructor
   private web3!: Web3;
   private abiDecoder: AbiDecoder;
