+++ packages/ethapi/src/helpers.ts
@@ -1,8 +1,33 @@
 import BigNumber from "bignumber.js";
+import * as Debug from "debug";
 import { Observable } from "rxjs/Observable";
+import * as Web3 from "web3";
 import { EthAddress } from "../../typescript-types/build";
 import { CivilErrors, isDefined } from "../../utils/build/src";
 import { EthApi } from "./ethapi";
+
+const debug = Debug("civil:ethapi:helpers");
+
+const DEFAULT_HTTP_NODE = "http://localhost:8545";
+
+export function detectProvider(): Web3.Provider {
+  let provider: Web3.Provider;
+  // Try to use the window's injected provider
+  if (hasInjectedProvider()) {
+    const injectedWeb3: Web3 = (window as any).web3;
+    provider = injectedWeb3.currentProvider;
+    debug("Using injected web3 provider");
+  } else {
+    // TODO(ritave): Research using Infura
+    provider = new Web3.providers.HttpProvider(DEFAULT_HTTP_NODE);
+    debug("No web3 provider provided or found injected, defaulting to HttpProvider");
+  }
+  return provider;
+}
+
+export function hasInjectedProvider(): boolean {
+  return typeof window !== "undefined" && (window as any).web3 !== undefined;
+}
 
 export function requireAccount(ethApi: EthApi): Observable<EthAddress> {
   return ethApi.accountStream.first().map(account => {
@@ -22,7 +47,13 @@ export async function currentNetwork(ethApi: EthApi): Promise<number> {
 }
 
 export function toWei(value: string | BigNumber | number, unit: EthereumUnits | string): BigNumber {
-  const unitPower = new BigNumber(EthereumUnits[unit as any]);
+  let unitValue: string;
+  if (unit in EthereumUnits) {
+    unitValue = EthereumUnits[unit as any];
+  } else {
+    unitValue = unit;
+  }
+  const unitPower = new BigNumber(unitValue);
   return new BigNumber(value).times(unitPower);
 }
